requirements.txt
asgiref==3.6.0
Django==4.2.1
django-debug-toolbar==4.0.0
djangorestframework==3.14.0
mypy==1.2.0
mypy-extensions==1.0.0
psycopg==3.1.9
psycopg-binary==3.1.9
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
